AI: on github page, at the end of simulation: Unhandled Promise Rejection: Reason: TypeError: augmented_state_json_proxy.destroy is not a function     at Object.augmentState (https://tmt1611.github.io/geom/static/js/api.js:352:40)     at showStateAtIndex (https://tmt1611.github.io/geom/static/js/main.js:179:40)     at HTMLButtonElement.<anonymous> (https://tmt1611.github.io/geom/static/js/main.js:726:19)
note that the local server works fine.
